🔨 refactor: 1st level of internal core splitting
